+++ CHANGELOG.md
@@ -2,8 +2,9 @@ CHANGELOG
 ---------
 
 * 0.4.1 (xx-xx-2013)
-  
+
   * Add compatibility with PHP-FFMpeg 0.4.
+  * Fix temporary files management.
 
 * 0.4.0 (07-04-2013)